Split tests files
